+++ utils/Training.java
@@ -3,20 +3,16 @@ import ai.djl.ndarray.NDArray;
 import ai.djl.ndarray.NDList;
 import ai.djl.ndarray.NDManager;
 import ai.djl.ndarray.types.DataType;
-import ai.djl.nn.AbstractBlock;
-import ai.djl.nn.Parameter;
 import ai.djl.training.EasyTrain;
 import ai.djl.training.Trainer;
 import ai.djl.training.dataset.ArrayDataset;
 import ai.djl.training.dataset.Batch;
 import ai.djl.translate.TranslateException;
-import ai.djl.util.Pair;
 
 import java.io.IOException;
 import java.util.Map;
 import java.util.function.BinaryOperator;
 import java.util.function.UnaryOperator;
-
 
 class Training {
 
@@ -53,28 +49,6 @@ class Training {
             NDArray gradient = param.getGradient();
             gradient.attach(subManager);
             param.subi(gradient.mul(lr).div(batchSize));
-        }
-    }
-
-    /** Clip the gradient. */
-    public static void gradClipping(Object net, int theta, NDManager manager) {
-        double result = 0;
-        NDList params;
-        params = new NDList();
-        for (Pair<String, Parameter> pair : ((AbstractBlock) net).getParameters()) {
-            params.add(pair.getValue().getArray());
-        }
-        for (NDArray p : params) {
-            NDArray gradient = p.getGradient().stopGradient();
-            gradient.attach(manager);
-            result += gradient.pow(2).sum().getFloat();
-        }
-        double norm = Math.sqrt(result);
-        if (norm > theta) {
-            for (NDArray param : params) {
-                NDArray gradient = param.getGradient();
-                gradient.muli(theta / norm);
-            }
         }
     }
